[Java] Add AbstractMessageTransceiver base class to reduce code duplication
